Revive fixes
